pyproject.toml
[project]
name = "commandnote"
version = "0.1.0"
description = "Command note tool with tree structure management based on MVC architecture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pywebview>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["models", "controllers", "views"]
